Implement fixed_size_mask with std::bitset
Use std::bitset<N> as data member of mask<T, fixed_size<N>>. This should
be optimizable to perfect code for AVX512 and similar ISAs that use
bitmasks.

SSE and AVX require conversion between the register and memory format.
This will be interesting to teach compilers to recognize and optimize
away.

Signed-off-by: Matthias Kretz <kretz@kde.org>